Renamed all namespaces to domain
